URL_P-main/backend/requirements.txt
pandas
numpy
scikit-learn
xgboost
joblib
tldextract
validators
fastapi
uvicorn
python-dotenv
openai
pymongo
python-jose[cryptography]
argon2-cffi